# cognition/eval-engine/routes/eval_routes.py
"""
Evaluation Routes - Agent evaluation endpoints.
"""
import sys
import os
import uuid
import asyncio
from typing import Dict, Any
from fastapi import APIRouter, Depends, HTTPException, status
from datetime import datetime

sys.path.append(os.path.dirname(os.path.dirname(__file__)))
from models import EvaluationRequest, EvaluationResult
from workers import AgentEvaluationWorker
from baseline_manager import BaselineManager

# Import auth from shared
sys.path.append(os.path.join(os.path.dirname(__file__), '..', '..', 'shared'))
from auth import get_current_user, User, Role

router = APIRouter(prefix="/v1", tags=["evaluation"])


@router.post("/evaluate", response_model=Dict[str, Any])
async def evaluate_agent(
    request: EvaluationRequest,
    current_user: User = Depends(get_current_user)
):
    """
    Submit agent for evaluation across environments.

    Requires: developer or admin role.

    Returns:
        Evaluation results with metrics, regression detection, and baseline comparison
    """
    # TODO: Implement POST /evaluate
    # 1. Check user role (require developer or admin)
    #    if current_user.role not in [Role.DEVELOPER, Role.ADMIN]:
    #        raise HTTPException(status_code=403, detail="Insufficient permissions")
    #
    # 2. Generate job_id = f"eval-{uuid.uuid4()}"
    #
    # 3. Get dependencies from app.state:
    #    worker: AgentEvaluationWorker
    #    baseline_manager: BaselineManager
    #
    # 4. results = []
    #    for environment in request.environments:
    #        baseline = None
    #        if request.compare_to_baseline:
    #            baseline = await baseline_manager.get_baseline(request.agent_type, environment)
    #
    #        result = await worker.evaluate_agent_in_env(
    #            agent_type=request.agent_type,
    #            agent_state=request.agent_state,
    #            hyperparameters=request.hyperparameters,
    #            environment=environment,
    #            num_episodes=request.num_episodes if not request.quick_mode else 50,
    #            baseline=baseline
    #        )
    #        results.append(result)
    #
    # 5. Compute Nash score if multiple agents (future: multi-agent eval)
    #
    # 6. Return {
    #        "job_id": job_id,
    #        "agent_type": request.agent_type,
    #        "results": results,
    #        "completed_at": datetime.utcnow().isoformat()
    #    }
    pass


@router.get("/jobs/{job_id}", response_model=Dict[str, Any])
async def get_job_status(
    job_id: str,
    current_user: User = Depends(get_current_user)
):
    """
    Get evaluation job status and results.

    Requires: viewer or higher role.

    Note: Phase 13.2 only supports synchronous evaluation.
    This endpoint returns cached results if available, otherwise 404.
    """
    # TODO: Implement GET /jobs/{job_id}
    # For Phase 13.2: Synchronous evaluation only
    # In Phase 13.3: Add job queue with Redis
    # For now: raise HTTPException(status_code=404, detail="Job not found")
    pass